Reorganized user API
Made POST and DELETE one API. Moved login end point out of the users folder and removed the users folder, leaving only users.ts as the API endpoint. Fixed fetch requests to reflect the changes
