+++ pages/api/users.tsx
@@ -1,0 +1,91 @@
+import type { NextApiRequest, NextApiResponse } from "next";
+import clientPromise from "../../util/db";
+import crypto from "crypto";
+import User from "../../models/user";
+import parseCookie from "../../util/parseCookie";
+
+/* POST - Creates a new user and requires a request body containing an email, password, and username field
+ * DELETE - Deletes the user permanently. Requires the token that should be sent as a cookie
+ * There is not GET function. That is handled via the getUser(context) util function
+ */
+
+const handler = async (req: NextApiRequest, res: NextApiResponse) => {
+  // Connect to the "users" collection on the database
+  const client = await clientPromise;
+  const db = client.db(process.env.DB);
+  const users = db.collection("users");
+
+  let username = "";
+  let password = "";
+  let token = "";
+
+  switch (req.method) {
+    /* -------------------- GET ------------------------ */
+    case "GET":
+      // 405: METHOD NOT ALLOWED
+      res.status(405).redirect("/");
+      /* Getting the user is done via the util function `getUser(context)`
+       to be able to retrieve the user object with Next.js */
+      break;
+    /* -------------------- POST ------------------------ */
+    // Create new user only if using a POST request
+    case "POST":
+      username = req.body.user;
+      const email = req.body.email;
+      password = req.body.pass;
+
+      // Check to see if user already exists (username or email)
+      const userExists = await users.findOne({ username: username });
+      if (userExists) {
+        // 409: CONFLICT
+        res.status(409).redirect("/sign-up?error=Username-already-in-use");
+        break;
+      }
+      const emailExists = await users.findOne({ email: email });
+      if (emailExists) {
+        // 409: CONFLICT
+        res.status(409).redirect("/sign-up?error=Email-already-in-use");
+        break;
+      }
+
+      // Hash the password for storage on the database
+      let hashPassword = crypto.createHash("sha256").update(password).digest("hex");
+
+      // create a user object and post it to the database
+      const newUser = new User(username, email, hashPassword, new Date());
+      users.insertOne(newUser);
+
+      // 201: CREATED
+      res.status(201).redirect("/login");
+      break;
+    /* -------------------- DELETE ------------------------ */
+    case "DELETE":
+      // Extract authorization token from cookies
+      const cookie = req.headers.cookie || "";
+      const parsedCookies = parseCookie(cookie);
+      token = parsedCookies.token;
+
+      // If the token exists
+      if (token) {
+        const deleted = await users.deleteOne({ token: token });
+        if (deleted.deletedCount > 0) {
+          // If the user has been deleted, status 200
+          res.status(200).json({ success: true });
+          break;
+        } else {
+          // If nothing was deleted, status 500: Internal server error
+          res.status(500).json({ success: false });
+          break;
+        }
+      } else {
+        // If no token is found, error 401: Unauthorized
+        res.status(401).json({ sucess: false });
+        break;
+      }
+    default:
+      // 405: METHOD NOT ALLOWED
+      res.status(405).redirect("/");
+  }
+}
+
+export default handler;
